+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-08-17 09:58:48
-// Commit: test(state): add unit tests for state business logic
+// Updated at: 2026-08-18 20:57:51
+// Commit: fix(state): fix memory leak by cancelling state subscription
 
 class stateComponentState {
-  final String commitRef = '3665';
-  final String lastMessage = 'test(state): add unit tests for state business logic';
-  final DateTime timestamp = DateTime.parse('2026-08-17T09:58:48+05:00');
+  final String commitRef = '3720';
+  final String lastMessage = 'fix(state): fix memory leak by cancelling state subscription';
+  final DateTime timestamp = DateTime.parse('2026-08-18T20:57:51+05:00');
 }